Simple Digit Convertor source program
Most Important tip in this simple project is that when you trying to create a solution in VS2012 or older or newer, then you must Select the "Windows Form Application" with C# language.The next step is this that you design a form with two Text Box and a Button, and this program should be replaced with old source code.
Ok! Your program is ready to use! Only input the english digits in text box 1 and see the output in text box 2!

Allah bless and good luck!